Properly encapsulate ES6 type defs. in module
